Separate writeCSVFile
